[AWSX-678] Step1: restructure files in the lambda forwarder
