+++ aws/logs_monitoring/exceptions.py
@@ -1,0 +1,6 @@
+class RetriableException(Exception):
+    pass
+
+
+class ScrubbingException(Exception):
+    pass
