+++ aws/logs_monitoring/tests/test_lambda_function.py
@@ -26,51 +26,20 @@ env_patch = patch.dict(
     },
 )
 env_patch.start()
-from lambda_function import (
-    invoke_additional_target_lambdas,
-    extract_metric,
-    extract_host_from_cloudtrails,
-    extract_host_from_guardduty,
-    extract_host_from_route53,
-    extract_trace_payload,
-    enrich,
-    transform,
-    split,
-    extract_ddtags_from_message,
-)
+from lambda_function import invoke_additional_target_lambdas
+from enrichment import enrich
+from transformation import transform
+from splitting import split
 from parsing import parse, parse_event_type
 
 env_patch.stop()
 
 
-class TestExtractHostFromLogEvents(unittest.TestCase):
-    def test_parse_source_cloudtrail(self):
-        event = {
-            "ddsource": "cloudtrail",
-            "message": {
-                "userIdentity": {
-                    "arn": "arn:aws:sts::601427279990:assumed-role/gke-90725aa7-management/i-99999999"
-                }
-            },
-        }
-        extract_host_from_cloudtrails(event)
-        self.assertEqual(event["host"], "i-99999999")
-
-    def test_parse_source_guardduty(self):
-        event = {
-            "ddsource": "guardduty",
-            "detail": {"resource": {"instanceDetails": {"instanceId": "i-99999999"}}},
-        }
-        extract_host_from_guardduty(event)
-        self.assertEqual(event["host"], "i-99999999")
-
-    def test_parse_source_route53(self):
-        event = {
-            "ddsource": "route53",
-            "message": {"srcids": {"instance": "i-99999999"}},
-        }
-        extract_host_from_route53(event)
-        self.assertEqual(event["host"], "i-99999999")
+class Context:
+    function_version = 0
+    invoked_function_arn = "arn:aws:lambda:sa-east-1:601427279990:function:inferred-spans-python-dev-initsender"
+    function_name = "inferred-spans-python-dev-initsender"
+    memory_limit_in_mb = "10"
 
 
 class TestInvokeAdditionalTargetLambdas(unittest.TestCase):
@@ -100,38 +69,6 @@ class TestInvokeAdditionalTargetLambdas(unittest.TestCase):
         )
 
 
-class TestExtractMetric(unittest.TestCase):
-    def test_empty_event(self):
-        self.assertEqual(extract_metric({}), None)
-
-    def test_missing_keys(self):
-        self.assertEqual(extract_metric({"e": 0, "v": 1, "m": "foo"}), None)
-
-    def test_tags_instance(self):
-        self.assertEqual(extract_metric({"e": 0, "v": 1, "m": "foo", "t": 666}), None)
-
-    def test_value_instance(self):
-        self.assertEqual(extract_metric({"e": 0, "v": 1.1, "m": "foo", "t": []}), None)
-
-    def test_value_instance_float(self):
-        self.assertEqual(extract_metric({"e": 0, "v": None, "m": "foo", "t": []}), None)
-
-
-class Context:
-    function_version = 0
-    invoked_function_arn = "arn:aws:lambda:sa-east-1:601427279990:function:inferred-spans-python-dev-initsender"
-    function_name = "inferred-spans-python-dev-initsender"
-    memory_limit_in_mb = "10"
-
-
-def create_cloudwatch_log_event_from_data(data):
-    # CloudWatch log event data is a base64-encoded ZIP archive
-    # see https://docs.aws.amazon.com/lambda/latest/dg/services-cloudwatchlogs.html
-    gzipped_data = gzip.compress(bytes(data, encoding="utf-8"))
-    encoded_data = base64.b64encode(gzipped_data).decode("utf-8")
-    return encoded_data
-
-
 class TestLambdaFunctionEndToEnd(unittest.TestCase):
     @patch("enhanced_lambda_metrics.LambdaTagsCache.get_cache_from_s3")
     def test_datadog_forwarder(self, mock_get_s3_cache):
@@ -149,7 +86,9 @@ class TestLambdaFunctionEndToEnd(unittest.TestCase):
         )
         context = Context()
         input_data = self._get_input_data()
-        event = {"awslogs": {"data": create_cloudwatch_log_event_from_data(input_data)}}
+        event = {
+            "awslogs": {"data": self._create_cloudwatch_log_event_from_data(input_data)}
+        }
         os.environ["DD_FETCH_LAMBDA_TAGS"] = "True"
 
         event_type = parse_event_type(event)
@@ -206,7 +145,9 @@ class TestLambdaFunctionEndToEnd(unittest.TestCase):
         cw_logs_tags_get.return_value = ["service:log_group_service"]
         context = Context()
         input_data = self._get_input_data()
-        event = {"awslogs": {"data": create_cloudwatch_log_event_from_data(input_data)}}
+        event = {
+            "awslogs": {"data": self._create_cloudwatch_log_event_from_data(input_data)}
+        }
 
         normalized_events = parse(event, context)
         enriched_events = enrich(normalized_events)
@@ -225,7 +166,9 @@ class TestLambdaFunctionEndToEnd(unittest.TestCase):
         cw_logs_tags_get.return_value = ["service:log_group_service"]
         context = Context()
         input_data = self._get_input_data()
-        event = {"awslogs": {"data": create_cloudwatch_log_event_from_data(input_data)}}
+        event = {
+            "awslogs": {"data": self._create_cloudwatch_log_event_from_data(input_data)}
+        }
 
         normalized_events = parse(event, context)
         enriched_events = enrich(normalized_events)
@@ -246,7 +189,9 @@ class TestLambdaFunctionEndToEnd(unittest.TestCase):
 
         context = Context()
         input_data = self._get_input_data()
-        event = {"awslogs": {"data": create_cloudwatch_log_event_from_data(input_data)}}
+        event = {
+            "awslogs": {"data": self._create_cloudwatch_log_event_from_data(input_data)}
+        }
 
         normalized_events = parse(event, context)
         enriched_events = enrich(normalized_events)
@@ -268,7 +213,9 @@ class TestLambdaFunctionEndToEnd(unittest.TestCase):
 
         context = Context()
         input_data = self._get_input_data()
-        event = {"awslogs": {"data": create_cloudwatch_log_event_from_data(input_data)}}
+        event = {
+            "awslogs": {"data": self._create_cloudwatch_log_event_from_data(input_data)}
+        }
 
         normalized_events = parse(event, context)
         enriched_events = enrich(normalized_events)
@@ -291,134 +238,12 @@ class TestLambdaFunctionEndToEnd(unittest.TestCase):
 
         return input_data
 
-
-class TestLambdaFunctionExtractTracePayload(unittest.TestCase):
-    def test_extract_trace_payload_none_no_trace(self):
-        message_json = """{
-            "key": "value"
-        }"""
-        self.assertEqual(extract_trace_payload({"message": message_json}), None)
-
-    def test_extract_trace_payload_none_exception(self):
-        message_json = """{
-            "invalid_json"
-        }"""
-        self.assertEqual(extract_trace_payload({"message": message_json}), None)
-
-    def test_extract_trace_payload_unrelated_datadog_trace(self):
-        message_json = """{"traces":["I am a trace"]}"""
-        self.assertEqual(extract_trace_payload({"message": message_json}), None)
-
-    def test_extract_trace_payload_valid_trace(self):
-        message_json = """{"traces":[[{"trace_id":1234}]]}"""
-        tags_json = """["key0:value", "key1:value1"]"""
-        item = {
-            "message": '{"traces":[[{"trace_id":1234}]]}',
-            "tags": '["key0:value", "key1:value1"]',
-        }
-        self.assertEqual(
-            extract_trace_payload({"message": message_json, "ddtags": tags_json}), item
-        )
-
-
-class TestMergeMessageTags(unittest.TestCase):
-    message_tags = '{"ddtags":"service:my_application_service,custom_tag_1:value1"}'
-    custom_tags = "custom_tag_2:value2,service:my_custom_service"
-
-    def test_extract_ddtags_from_message_str(self):
-        event = {
-            "message": self.message_tags,
-            "ddtags": self.custom_tags,
-            "service": "my_service",
-        }
-
-        extract_ddtags_from_message(event)
-
-        self.assertEqual(
-            event["ddtags"],
-            "custom_tag_2:value2,service:my_application_service,custom_tag_1:value1",
-        )
-        self.assertEqual(
-            event["service"],
-            "my_application_service",
-        )
-
-    def test_extract_ddtags_from_message_dict(self):
-        loaded_message_tags = json.loads(self.message_tags)
-        event = {
-            "message": loaded_message_tags,
-            "ddtags": self.custom_tags,
-            "service": "my_service",
-        }
-
-        extract_ddtags_from_message(event)
-
-        self.assertEqual(
-            event["ddtags"],
-            "custom_tag_2:value2,service:my_application_service,custom_tag_1:value1",
-        )
-        self.assertEqual(
-            event["service"],
-            "my_application_service",
-        )
-
-    def test_extract_ddtags_from_message_service_tag_setting(self):
-        loaded_message_tags = json.loads(self.message_tags)
-        loaded_message_tags["ddtags"] = ",".join(
-            [
-                tag
-                for tag in loaded_message_tags["ddtags"].split(",")
-                if not tag.startswith("service:")
-            ]
-        )
-        event = {
-            "message": loaded_message_tags,
-            "ddtags": self.custom_tags,
-            "service": "my_custom_service",
-        }
-
-        extract_ddtags_from_message(event)
-
-        self.assertEqual(
-            event["ddtags"],
-            "custom_tag_2:value2,service:my_custom_service,custom_tag_1:value1",
-        )
-        self.assertEqual(
-            event["service"],
-            "my_custom_service",
-        )
-
-    def test_extract_ddtags_from_message_multiple_service_tag_values(self):
-        custom_tags = self.custom_tags + ",service:my_custom_service_2"
-        event = {"message": self.message_tags, "ddtags": custom_tags}
-
-        extract_ddtags_from_message(event)
-
-        self.assertEqual(
-            event["ddtags"],
-            "custom_tag_2:value2,service:my_application_service,custom_tag_1:value1",
-        )
-        self.assertEqual(
-            event["service"],
-            "my_application_service",
-        )
-
-    def test_extract_ddtags_from_message_multiple_values_tag(self):
-        loaded_message_tags = json.loads(self.message_tags)
-        loaded_message_tags["ddtags"] += ",custom_tag_3:value4"
-        custom_tags = self.custom_tags + ",custom_tag_3:value3"
-        event = {"message": loaded_message_tags, "ddtags": custom_tags}
-
-        extract_ddtags_from_message(event)
-
-        self.assertEqual(
-            event["ddtags"],
-            "custom_tag_2:value2,custom_tag_3:value3,service:my_application_service,custom_tag_1:value1,custom_tag_3:value4",
-        )
-        self.assertEqual(
-            event["service"],
-            "my_application_service",
-        )
+    def _create_cloudwatch_log_event_from_data(self, data):
+        # CloudWatch log event data is a base64-encoded ZIP archive
+        # see https://docs.aws.amazon.com/lambda/latest/dg/services-cloudwatchlogs.html
+        gzipped_data = gzip.compress(bytes(data, encoding="utf-8"))
+        encoded_data = base64.b64encode(gzipped_data).decode("utf-8")
+        return encoded_data
 
 
 if __name__ == "__main__":
